pyproject.toml

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "feduciary"
version = "0.0.1"
authors = [
  { name="Asim Ukaye", email="asim.ukaye@mbzuai.ac.ae" },
]
description = "A fair federated learning package for PyTorch."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD-3-Clause License",
    "Operating System :: OS Independent",
]

dependencies = [
    "torch>=2.0.1",
    "torchvision>=0.15.0",
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.4.0",
    "pandas>=2.0.3",
    "wandb>=0.15.0",
    "tqdm>=4.64.0",
    "flwr>=1.6.0",
    "icecream>=2.1.1",
    "opencv-python>=4.9.0",
    "Pillow>=9.3.0",
    "tensorboard>=2.13.0",
    "torchtext>=0.15.2",
    "torchdata>=0.6.1",
    "transformers>=4.31.0",
    "seaborn>=0.13.1",
    "hydra-core>=1.3.2",
    "albumentations>=1.3.0",
]

[project.urls]
Homepage = "https://github.com/asimukaye/feduciary"
Issues = "https://github.com/asimukaye/feduciary/issues"